thirdeye-py-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
strawberry-graphql[fastapi]==0.235.0
aiodataloader==0.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
SQLAlchemy==2.0.25
aiomysql==0.2.0
alembic==1.13.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
tenacity==8.2.3
prometheus-client==0.19.0
loguru==0.7.2
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
mypy==1.8.0
types-PyYAML==6.0.12.12
PyYAML==6.0.1
